User login and sign up, cancelReservation

+++ src/main/java/application/service/ReservationService.java
@@ -1,6 +1,9 @@
 package application.service;
 
 import application.model.Reservation;
+import application.model.RestaurantTable;
+import application.model.User;
+import application.model.util.DateTimeSlot;
 import application.repository.DateTimeSlotRepository;
 import application.repository.ReservationRepository;
 import application.repository.RestaurantTableRepository;
@@ -8,6 +11,9 @@ import application.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import javax.transaction.Transactional;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.LocalTime;
 import java.util.List;
 
 @Service
@@ -25,12 +31,42 @@ public class ReservationService {
     @Autowired
     private DateTimeSlotRepository dateTimeSlotRepository;
 
+
     @Transactional
-    public Reservation reserveTable(Reservation reservation) {
-        if (tableRepository.existsById(reservation.getRestaurantTable().getId()) && userRepository.existsById(reservation.getUser().getId())) {
-            reservationRepository.save(reservation);
+    public User isAuthorized(String authToken) {
+        return userRepository.findByAuthToken(authToken);
+    }
+
+    @Transactional
+    public Reservation reserveTable(User user, long tableId, String date, List<Double> timeSlot) {
+        Reservation reservation = new Reservation();
+        RestaurantTable table = tableRepository.getById(tableId);
+        if (table == null) {
+            return null;
         }
-        return null;
+        reservation.setRestaurantTable(table);
+        DateTimeSlot dateTimeSlot = DateTimeSlot.convertToDateTimeSlot(date, timeSlot.get(0), timeSlot.get(1));
+        if (dateTimeSlot == null) {
+            return null;
+        }
+        reservation.setDateTimeSlot(dateTimeSlot);
+        reservation.setUser(user);
+        reservation.setConfirmed(false);
+        if (!hasFreeTimeSlot(table, dateTimeSlot)) {
+            return null;
+        }
+        return reservationRepository.save(reservation);
+    }
+
+    private boolean hasFreeTimeSlot(RestaurantTable table, DateTimeSlot freeTimeSlot) {
+        List<Reservation> reservationsForSpecifiedDate = table.getReservations().stream().filter(reservation -> reservation.getDateTimeSlot().getDate().equals(freeTimeSlot.getDate())).toList();
+        // detect potential collision
+        for (Reservation reservation : reservationsForSpecifiedDate) {
+            if (reservation.getDateTimeSlot().isCollision(freeTimeSlot)) {
+                return false;
+            }
+        }
+        return true;
     }
 
     @Transactional
@@ -41,8 +77,22 @@ public class ReservationService {
     }
 
     @Transactional
-    public void cancelReservation(Long id) {
+    public boolean cancelReservation(User user, Long id) {
+        Reservation reservation = reservationRepository.getById(id);
+        if (!user.getReservations().contains(reservation)) {
+            return false;
+        }
+        if (reservation.getConfirmed()) {
+            return false;
+        }
+        LocalDateTime currentTime = LocalDateTime.now();
+        LocalTime reservationStartTime = reservation.getDateTimeSlot().getStartTime();
+        LocalDate reservationDate = reservation.getDateTimeSlot().getDate();
+        if (currentTime.plusHours(12).compareTo(LocalDateTime.of(reservationDate, reservationStartTime)) > 0) {
+            return false;
+        }
         reservationRepository.deleteById(id);
+        return true;
     }
 
     @Transactional
